package: move the one spec to be an EL7.2 "example"
Since there could be different spec files for different distros,
move the one that was provided to examples/powerman_el72.spec to
be clear that it is just an example and that distro packagers could
use as a starting point for their own specs.

That said, try to keep the example functional for EL7.