Add --with-group= option to config.ac
The powerman.service file has configuration options to run the daemon
as a specific user and group. The current default for both is "daemon".
Changing the user is possible through --with-user= configure option,
but there is no option for changing the group.

In order to allow the group value to be configured at build time, a
configure option similar to the --with-user= option should be provided.

This change creates the --with-group= option, setting the
@RUN_AS_GROUP@ macro for autoconf files.