Allow configuration of service user and group
The powerman systemd service file currently sets the running user and
group statically to "daemon". Certain systems (SUSE) consider grouping
all daemons under one general user and group a security concern. The
accepted policy is to run each daemon as it's own user and group in
order to contain potential security violations.

This fix puts the powerman.service file under the control of the
configure script, and allows changing the service user and group using
the --with-user= and --with-group= options respectively.