drop unused top level meta files
Problem: README, ChangeLog, INSTALL, AUTHORS, and TODO files
are outdated/superfluous.

Set automake foreign mode so that the project does not have to adhere
to the GNU project template.

Drop README and TODO  which are outdated.
Drop ChangeLog which is empty.
Drop AUTHORS which duplicates info in the git commit log.
Drop INSTALL which is gnu boilerplate that nobody reads.