build: conditional use of gcc warning flags is broken
Problem: @GCCWARN@ is substituted even when the compiler is clang.

Import the AX_COMPILER_VENDOR macro and use it to test whether the
compiler vendor is clang or gnu.

Change the warning flags macro name to WARNING_CFLAGS, then conditionally
set it to different values for clang vs gnu.  Tell clang not to complain
about unknown-warning-option errors.  Although this is not currently an
issue in powerman, it is in other projects, so might as well.