systemd: don't include systemd configs in dist
Problem: powerman.service is now generated by configure, so
it should not be included in the distribution.

Use systemdsystemunit_DATA instead of _SCRIPTS and drop from
EXTRA_DIST (which was redundant anyway).

Similarly, use _DATA instead of _SCRIPTS for the tmpdir.d
config file, since it is also generated by configure.