build: use libtool for convenience libraries
Problem: libtool is used to build the installed libpowerman
library, but not convenience libraries which could lead to
problems.

Right now the API is disjoint from the rest of powerman but
if it were to use convenience libraries, we could run into
trouble with non-PIC code.

Use libtool for the whole project.