etc: set timeout in redfishpower devices
Problem: If a user modifies the powerman timeout in a redfishpower
device file, they may not be aware that it should also be modified
in redfishpower.

Add a call to settimeout in the login script of redfishpower devices.
This can alert users that any changes to the powerman timeout should
also be adjusted in redfishpower.