redfishpower: support setstatuspollinginterval
Problem: The wait until on/off polling interval in redfishpower
is currently hard coded to 1 second.  On some slow systems, it has
been observed that on/off can take upwards of a minute.  On that
hardware, a 1 second retry delay can lead to an excess amount of
traffic on the network.

Support a prompt command that will allow change of the status polling
interval.  Document new command in redfishpower(8).  Update redfishpower
test helper to parse setstatuspollinginterval.