etc: set polling interval in redfishpower devices
Problem: The status polling interval can too short on systems
where redfishpower is known to be slow.  This can lead to an excessive
number of status messages being sent out.  The new
setstatuspollinginterval command can allow users to configure
this interval in redfishpower device files, but most users will not
know it exists.

Add a call to setstatuspollinginterval in the login script of
redfishpower devices.  This can alert users that this configuration
exists.