man: add redfishpower modification info
Problem: Due to the nature of redfishpower being a REST interface,
many users are copying and pasting the redfishpower device files
and slightly tweaking them for their systems.  Users could use
more information on how to do this.  In addition, there are potential
pitfalls on other systems that users need to be aware of.

Add a section on modifications to redfishpower(8).  Add pointers to
this section in all redfishpower device files.

Fixes #67