package: set SHELL in powerman's environment
Powerman may need to run things to connect to power control devices,
such as an ssh command.  For ssh to operate properly it needs to have
SHELL set to a valid shell in it's environment.

Note from @garlick: this was submitted in 2017 as pr #24 but was
left unmerged due to pending git machinations.  Let's just pull it in.