powerman: add --retry-connect=N option for test
Problem: there is no way to ask the powerman client to retry
a connection attempt when starting a server in the background
in test.

The server may take an indeterminate amount of time to become ready.

Add -R,--retry-connect=N which causes the client handle a connect
failure differently.  Instead of exiting immediately, sleep 1s and
retry the connection attempt, up to N times.

This should be considered a test-only option, thus it doesn't appear
in the help output or man page.