testsuite: convert remote powerman test to sharness
Problem: test/t35 and t36 cover using a remote powermand
as a power controller using old test infrastructure.

Convert t35 to sharness.
Don't bother converting t36 since all it does is rename the plugs,
which is a general powerman feature not related to this device.

Remove the old test scripts and data.