systemd: run as Type=simple
Problem: powermand runs as a Type=forking systemd service,
but Type=simple is recommended by systemd.service(5).

Switch to Type=simple (the default).

Drop the tmpfiles.d configuration which was for managing
/run/powerman on systemd < v233 (now ancient).  Powerman doesn't
need the rundir anyway when not running in daemon mode, since all
it does is store a pidfile there.

Fixes #13