redfishpower: remove --hostsfile option
Problem: The redfishpower --hostsfile option is oddly included in the
project as an option.  The suspicion is it was copy and paste from
another project.  The short option for the option is not configured,
but the long option is.  It is not documented in --help output, but is
oddly documented in the manpage.

Drop the option.