README: update supported protocols
Problem: README mentions X10 and VXI-11 but not Redfish

VXI-11 (and GPIB) support is based on an abandoned gpib-utils
project. The dev scripts remain but we just dropped testing for it.
Let's not advertise.

X10 comes with the Insteon support in plmpower, but it doesn't really
have complete powerman support because it doesn't support a query.
Let's not advertise that one either.

Redfish is actually being used in LLNL clusters, so mention that.