redfishpower: move initialization into function
Problem: The "hosts" data structure is initialized if the user
specifies "-h" on the command line.  Which this is ok for this
one data structure, but it can get annoying if additional future
initializations are done the same way.

Create a new init_redfishpower() function to initializes redfishpower
libs and data structures.  Adjust some follow on logic accordingly.