send response after powermand --stdio hangup
Problem: powermand --stdio exits immediately when stdin is
closed, dropping any in-flight response.

It is convenient in testing to be able to pipe commands to
powermand in this mode, but we can't check the result or be
assured the command completed.

Don't exit until both stdin has EOF and any pending command
has completed.

Fixes #130