powermand: fix file descriptor leak
Problem: in the recent conversion of coprocesses from forkpty(3) to
socketpair(2), the unused side of the pair is not closed in the
parent, which causes file descriptors to leak as coprocesses are
restarted on error.

Close the file descriptor in the parent.

Close its counterpart in the child (just good hygiene, not a leak).