powerman: log onoff messages to stderr not syslog
Problem: when we run the test suite, on/off messages appear
in syslog.

There is no longer a need to log anything directly to syslog.
When run under systemd, the systemd journal captures stderr
(and routes to syslog if so configured).

Log those messages to stderr.
Use a <level> prefix to preserve the ability to set the severity
of those messages in the config file.

Fixes #138