redfishpower: group curl code together
Problem: In several functions curl API calls are not "grouped"
together.  They are oddly called at different points in the
function even though it is not necessary.  This is likely due to
early development and cleanup was forgotten.

Group the curl API calls together so the code is a little cleaner.