redfishpower: support setpath configuration
Problem: Currently in redfishpower, a single global setting for
stat, on, and off paths is configured for all plugs.  This
may not be convenient for some environments, where some specific
plugs need to have different paths for power operations.

Support a new configuration "setpath" that can set paths for
specific plugs.  The remaining configurations remain in place
to configure the "defaults".