redfishpower: use terms consistently
Problem: The following terms are used interchangeably within the
redfishpower code:

path & url
data & postdata

Solution: Consistently use the term "path" to describe "paths" input by the user
for configuration and use the term "postdata" for http post data.  The term
"url" should only be used when the full URL is constructed (i.e. when
http:// is prefixed).

Update terms in the manpage as well.