redfishpower: support setplugs config
Problem: Redfishpower currently assumes plugnames are equal
to the hosts configured on the redfishpower command line.  In the
future this can lead to difficulty doing advanced configuration.
A different device file would be needed for any hosts that are
different.

Solution: Support a new setplug config which allows users to map
logical plugs to the indicies of the hosts configured.  For example:

setplugs Node[0-15] [0-15]

This creates a set of plugs named Node[0-15].  These may be used
in place of nodenames input on the command line.