redfishpower: check hostname legality later
Problem: An internal function called parse_input_hosts() checks if
all the inputs for a power command match hosts configured by the
user.  In the future, plug names will be used instead, so the check
will have no use in the future.

Solution: Remove parse_input_hosts() and check that the inputs are
legal as the input targets are iterated.