redfishpower: make cmd lists globals
Problem: The activecmds and delayedcmds lists are passed around
in a lot of places.  In the future additional lists will need to be
passed around to functions.  There is only one each of these lists, so it
serves no purpose to constantly pass them around everywhere.

Make the activecmds and delayedcmds lists globals.