redfishpower: support parent/hierarchy of plugs
Problem: In some hardware, there is a hierarchy for power control.  For
example, In bladed systems the chassis must be powered on before
blades can be powered on.  Without knowledge of this dependency, power
control can become difficult to use.

Solution: Support configuration of hierarchies through an optional "parent"
configuration in the "setplugs" option.

With knowledge of hierarchy, support following behavior semantics in
redfishpower.

When doing a power operation, check parents/ancestors power status first.

If ancestors are all on, then redfishpower can perform power
on/off/status on the target.

If ancestor is off, power status is defined as off for all
descendants.  Power on cannot be done.  Power off is assumed to succeed.

As special case, if powering on both ancestors and children, power on
will not be allowed.  Testing has shown that delays are necessary between
levels due to hardware "initialization" going on in the background.  This
delay is highly variable and behavior was not consistent.

Powering off both ancestors and children is allowed,  descendants
will wait until ancestor power offs are complete.  Then by definition,
descendants are then assumed to be off.

If any ancestors have status unknown or get an error, that status carries to
descendants and results in errors.