redfishpower: send http request after cmd active
Problem: Whenever a power command message struct is created,
it is immediately initialized within curl, which means curl will
immediately issue the http request once we enter the primary
loop.  This isn't what is desired for "delayed cmds", i.e. status
polling that we want to poll at a later time.

This happens to work right now because the only delayed commands
we do is status polling.  If the current status polling
happens to get the wrong answer, no biggie, we'll try again a
second later when we issue the next status poll.  In the future,
this will not be the case.  We will want delayed commands to issue
their http requests at a later time.

Only initialize curl on a power command once it is added to the
active list.