etc: do not use native cycle with redfishpower
Problem: A native power cycle with the redfish protocol is
inherently racy.  After performing a power cycle and immediately
getting the power status, there is no way to know if the power
status of "on" is from before or after the cycle.  This can be
confusing to users.

In all redfishpower devices files, define power cycles as an
off, delay, and on.  Do not use native redfish cycle/restart.

Fixes #148