etc: add delay in redfishpower cycle
Problem: When a power cycle is implemented as an off followed
by an on, a delay between the two is typically added in case the
system needs to "settle" a bit.

This is not strictly necessary for redfishpower, since it verifies
the power is off before performing the on.  However, since it is
the common pattern in powerman, it couldn't hurt to add a small
delay just in case the system needs "settling" time.

Add a small 2 second delay in cycle_ranged in
etc/devices/redfishpower-cray-windom.dev.