powerman: do not always call "all" query script
Problem: When selecting a script to use for an action, the "all"
version of a script is always used if the action is a query action
(e.g. power status query, beacon query, etc.).

This can be problematic in situations where blades/nodes/etc. are
not fully populated, such as in a chassis.  We do not want to query
"all" things in that case.

Solution: Only call the all query script if all plugs are requested
OR the singlet version of the script is not defined."?  For device files
where unpopulated plugs may exist, it can define an all script and singlet
script for each scenario.