redfishpower: add more debug/verbose info
Problem: When powering on a target, it will sometimes enter the
"PoweringOn" state but regress back to the "off" state.  This is
likely due to a hardware problem.  It can be hard to diagnose when
the only clue is a power on has "timed out".

Add some additional output to help diagnose this situation.