redfishpower: adapt status polling interval
Problem: The status polling interval is hard coded to 1 second long.
This can result in an excessive number of polling messages being sent
when it is known that some hardware takes 20-50 seconds to complete
a power operation.

Solution: Support a modified "eponential backoff" modification of the
status polling interval.  The modified algorithm is based on observations
of how long it typically takes to complete power operations on hardware.
The status polling interval begins at one second, but it gets capped
at 4 seconds.