redfishpower: support plug parents
Problem: Some chassis environments have a power control
hierarchy, where some components must be powered on before
the "child" components can be powered on.  There is no
infrastructure for this in the plug module.

Add support of a "parent" within each plug and a collection
of helper functions to navigate the resulting hierarchy
that can be configured within plugs.

Update callers to plugs_add() accordingly.