powerman: support new --diag option
Problem: When power control/query to a target fails, there is no way for
a user to know why it failed.

Add a new --diag to powerman that will inform powermand to send
diagnostic information about why a power operation failed.
This only works with setplugstate and the new setresult statement.