powerman: use singlet script if targeting one plug
Problem: When selecting a script to use for an action (i.e.  power on)
powerman may select the "all" or "ranged" script if they are available.
This is done even if a single plug is targeted.  There is nothing wrong
with this, but may not be what users expect.

If the user is specifically targeting only one plug, select the
singlet script over any other script.