t: add --bad-plug option to vpcd
Problem: Sometimes it would be useful to have errors reported
from vpcd, but there is no way to do that.

Solution: Add a --bad-plug option to vpcd.  It will ensure that
power operations to that specific plug never work.  Update vpc.dev
to handle the new possible output of "ERROR".