powerman: support error diagnostics with setresult
Problem: When power control to a target fails, there is no way for
a user to know why it failed except through the very verbose
--telemetry output.

When `setresult` is used in a device script, send text indicating
why the power operation failed to a specific plug.  From the client side,
have this be output to stderr.