redfishpower: output more detailed error messages
Problem: Redfishpower decided to output a generic "network error"
message when either a connection error occurred (i.e. typically
if the host is down) vs a operation timeout (i.e. such as a failed
DNS lookup).  All other errors are generically output as "error".
It would be useful to know which one actually happened.

Call curl_easy_strerror() on the error, to get the appropriate
error message for the error code.