redfishpower: support --resolve-hosts workaround
Problem: A DNS race in libcurl exists in versions before 7.66.  This can
lead to redfishpower returning "Timeout was reached" for power operations.

Support a new --resolve-hosts option which will resolve the inputed
hostnames, and then pass those IP addresses to libcurl instead of the host.
This appears to work around the race.

Fixes #184