powerman: correct potential segfault
Problem: In _process_setresult() if the plug name is not found
in the inputted arglist, it can result in a segfault.  This
can occur when power operations on dependent targets (e.g. the
parent of node needs to be powered on) have errors, leading to
an unexpected host having a "power result".

Check that arg is non-NULL before trying to dereference it.

Fixes #197