testsuite: skip el capitan test by default
Problem: the debian riscv64 build environment is failing on the
el capitan scaling test.

We don't have detailed test output to diagnose the failure but we
can probably assume that this is a fairly under-powered builder.

Set the EXPENSIVE built-in test prereq so that this test only runs
if TEST_LONG is set in the environment or the --long test option is
given.

To avoid losing coverage in github CI, also allow the test to run
if GITHUB_CI is set in the environment, and then set it where the
workflow runs make check.

Fixes #206

more